requirements.txt
pytest==7.4.3
pytest-html-reporter==0.2.9
selenium==4.16.0
pyyaml